ingestion/producer/requirements.txt
kafka-python==2.0.2
faker==19.6.2
python-dotenv==1.0.0
requests==2.31.0